origin_source_code/requirements.txt
allure-pytest==2.8.40
allure-python-commons==2.8.40
aniso8601==9.0.1
atomicwrites==1.4.0
attrs==20.3.0
bcrypt==3.2.0
by==0.0.2
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
cryptography==3.4.7
decorator==5.0.7
Deprecated==1.2.13
dictdiffer==0.9.0
dpkt==1.9.6
flasgger==0.9.5
Flask==2.0.3
flask-apidoc==1.2.0
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-Swag==0.1.2
flask-swagger==0.2.14
idna==2.10
importlib-metadata==4.0.1
inflection==0.5.1
iniconfig==1.1.1
isodate==0.6.0
JPype1==1.2.1
jsonpath==0.82
jsonpath-rw==1.4.0
jsonschema==3.2.0
MarkupSafe==2.0.1
marshmallow==3.13.0
mistune==0.8.4
Naked==0.1.31
numpy==1.19.5
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
packaging==20.9
paramiko==2.7.2
pluggy==0.13.1
ply==3.11
py==1.10.0
pycparser==2.20
pycryptodome==3.10.1
PyMySQL==1.0.2
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
redis==4.0.2
requests==2.25.1
schedule==1.1.0
selenium==3.141.0
shellescape==3.8.1
six==1.15.0
sshtunnel==0.4.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
wrapt==1.13.3
zipp==3.4.1
nacos-sdk-python==0.1.12